Split out TagProducer.
