+++ source/common/config/utility.cc
@@ -15,6 +15,7 @@
 #include "common/json/config_schemas.h"
 #include "common/protobuf/protobuf.h"
 #include "common/protobuf/utility.h"
+#include "common/stats/tag_producer_impl.h"
 #include "common/stats/stats_impl.h"
 
 namespace Envoy {
